refactor(core): TileTree seam — generalize traversal over a tile source
Introduce source::{TileId, TileProperties, TileTree}: TileId becomes an
opaque u64 handle (arena index for Tileset, encoded (z,x,y) for terrain).
The SSE traversal now walks a &dyn TileTree (roots/children/properties)
instead of a concrete Tileset, so it can drive 3D Tiles and terrain with the
same code. Tileset implements TileTree; multiple roots supported; depth is
tracked through the recursion. Zero behavior change — all tests green.

+++ crates/tuile-core/src/cache.rs
@@ -6,7 +6,7 @@
 //! Invariant: a currently selected tile is never evicted — the budget may
 //! be temporarily exceeded rather than dropping visible content.
 
-use crate::tileset::TileId;
+use crate::source::TileId;
 use std::collections::{HashMap, HashSet};
 
 #[derive(Debug, Clone, Copy)]
@@ -98,7 +98,7 @@ impl ResidentCache {
 mod tests {
     use super::*;
 
-    fn t(n: u32) -> TileId {
+    fn t(n: u64) -> TileId {
         TileId(n)
     }
